Add IsSupported() API which also parses /proc/self/status
We can use `/proc/self/status` to pre-check if seccomp is available on the
local system. This also means that we can call this function without
having a need for the `seccomp` build tag. If the build tag is available,
then the `IsEnabled()` function will now re-use the new `IsSupported()`
API.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>